Move platform independent functionalities interfaces
