+++ cannons-bukkit/src/main/java/at/pavlov/cannons/interfaces/holders/AimingDataHolder.java
@@ -1,7 +1,7 @@
 package at.pavlov.cannons.interfaces.holders;
 
 import at.pavlov.internal.cannon.data.AimingData;
-import at.pavlov.cannons.interfaces.functionalities.Updatable;
+import at.pavlov.internal.cannon.functionalities.Updatable;
 
 public interface AimingDataHolder extends Updatable {
     AimingData getAimingData();
