+++ cannons-bukkit/src/main/java/at/pavlov/cannons/interfaces/holders/CannonPositionHolder.java
@@ -2,15 +2,13 @@ package at.pavlov.cannons.interfaces.holders;
 
 import at.pavlov.cannons.Cannons;
 import at.pavlov.cannons.cannon.data.CannonPosition;
-import at.pavlov.cannons.interfaces.functionalities.Updatable;
+import at.pavlov.internal.cannon.functionalities.Updatable;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.World;
-import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
 import org.bukkit.util.Vector;
 
-import java.util.List;
 import java.util.UUID;
 
 public interface CannonPositionHolder extends Updatable {
